add postman to bruno translation
