+++ packages/bruno-converters/src/postman/postman-translations.js
@@ -15,11 +15,17 @@ const replacements = {
   'pm\\.expect\\(': 'expect(',
   'pm\\.environment\\.has\\(([^)]+)\\)': 'bru.getEnvVar($1) !== undefined && bru.getEnvVar($1) !== null',
   'pm\\.response\\.code': 'res.getStatus()',
-  'pm\\.response\\.text\\(': 'res.getBody()?.toString(',
+  'pm\\.response\\.text\\(\\)': 'JSON.stringify(res.getBody())',
   'pm\\.expect\\.fail\\(': 'expect.fail(',
   'pm\\.response\\.responseTime': 'res.getResponseTime()',
   'pm\\.environment\\.name': 'bru.getEnvName()',
+  'pm\\.response\\.status': 'res.statusText',
+  'pm\\.response\\.headers': 'req.getHeaders()',
   "tests\\['([^']+)'\\]\\s*=\\s*([^;]+);": 'test("$1", function() { expect(Boolean($2)).to.be.true; });',
+  'pm\\.request\\.url': 'req.getUrl()',
+  'pm\\.request\\.method': 'req.getMethod()',
+  'pm\\.request\\.headers': 'req.getHeaders()',
+  'pm\\.request\\.body': 'req.getBody()',
   // deprecated translations
   'postman\\.setEnvironmentVariable\\(': 'bru.setEnvVar(',
   'postman\\.getEnvironmentVariable\\(': 'bru.getEnvVar(',
